logratio-normal-multinomial random sample

logratio-normal-multinomial random sample

rlrnm(n = NULL, size, mu, sigma, B = NULL, probs = FALSE)

Arguments

n

sample size. if ommited, the value is obtained as the length of vector size (in that case, all parameters should be called using their name).

size

vector to set the multinomial sampling size. vector is reused to have length equal parameter n

mu

logratio-normal-multinomial mu parameter

sigma

logratio-normal-multinomial mu parameter

B

basis in which the parameters are expressed with

probs

logical indicating whether multinomial probabilities should be returned

Value

logratio-normal-multinomial random sample

Examples

mu = c(0,1)
sigma = diag(2)
rlrnm(10, 1000, mu, sigma)
#>       [,1] [,2] [,3]
#>  [1,]  221  683   96
#>  [2,]  209  601  190
#>  [3,]  532  294  174
#>  [4,]  407  577   16
#>  [5,]  240  587  173
#>  [6,]  585  398   17
#>  [7,]  361  626   13
#>  [8,]  244  735   21
#>  [9,]  190  299  511
#> [10,]  105  422  473