Multinomial random sample

Multinomial random sample

rmultinomial(n = NULL, size, p)

Arguments

n
number of random samples. Ignored if p is a matrix

size
vector to set the multinomial sampling size. vector is reused to have length equal to the number of rows of probs

p
vector or matrix specifiying the probability of each class. If a matrix is provided the number of rows will be equal to the number of samples to generate. Row number i defines the multinomial probability to be used when generating sample i. In this last scenario, parameter names (size and p) should be provided when calling the function.

Value

multinomial random sample

Examples

rmultinomial(10, 4, c(0.3, 0.4, 0.3))

#>       [,1] [,2] [,3]
#>  [1,]    2    2    0
#>  [2,]    3    1    0
#>  [3,]    3    0    1
#>  [4,]    2    1    1
#>  [5,]    2    1    1
#>  [6,]    1    3    0
#>  [7,]    1    1    2
#>  [8,]    0    1    3
#>  [9,]    3    1    0
#> [10,]    1    3    0

probs = matrix(c(0.2, 0.4, 0.4,
                 0.5, 0.1, 0.4), byrow = TRUE, nrow = 6, ncol = 3)
size = c(10, 500, 10000)
rmultinomial(size = size, p = probs)

#>      [,1] [,2] [,3]
#> [1,]    2    4    4
#> [2,]  269   62  169
#> [3,] 1993 3985 4022
#> [4,]    4    2    4
#> [5,]   95  197  208
#> [6,] 5011 1038 3951